pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "activity-agent"
version = "0.1.0"
description = "Local PC activity monitor: screenshots, OCR, LLM, SQLite"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10,<3.14"
dependencies = [
    "mss>=9.0.0",
    "pywin32>=306; sys_platform == 'win32'",
    "requests>=2.28.0",
    "paddlepaddle>=3.3.0,<3.4.0",
    "paddleocr>=3.0.0,<3.5.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
activity-agent-web = "activity_agent.ui.web_app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"activity_agent.ui" = ["templates/*.html", "static/*.css"]
